+++ vsl/atom.py
@@ -1,6 +1,6 @@
 import pygame
 
-from vsl import system, typing
+from vsl import sys, typing
 
 from vsl.utils import to_absolute, to_absolute_points
 
@@ -10,7 +10,7 @@ def rectangle(relative_x, relative_y, relative_width, relative_height, color):
     position = to_absolute((relative_x, relative_y))
     size = (to_absolute(relative_width), to_absolute(relative_height))
     rect = (position, size)
-    return pygame.draw.rect(system.screen, color, rect)
+    return pygame.draw.rect(sys.screen, color, rect)
 
 
 def circle(
@@ -22,7 +22,7 @@ def circle(
     position = to_absolute((relative_x, relative_y))
     radius = to_absolute(relative_radius)
     color = typing.color(color)
-    pygame.draw.circle(system.screen, color, position, radius)
+    pygame.draw.circle(sys.screen, color, position, radius)
 
 
 def polygon(
@@ -33,5 +33,5 @@ def polygon(
     ):
     points = to_absolute_points(relative_points)
     color = typing.color(color)
-    pygame.draw.polygon(system.screen, color, points, 0 if fill else thickness)
+    pygame.draw.polygon(sys.screen, color, points, 0 if fill else thickness)
