+++ vsl/layout.py
@@ -1,4 +1,4 @@
-from vsl import atom, system
+from vsl import atom, sys
 
 from vsl.color import *
 
@@ -8,7 +8,7 @@ def grid(
         nb_rows=2,
         cell=lambda row, col: atom.rectangle((0.1, 0.1), 0.9, 0.9, WHITE)
     ):
-    full_width, full_height = system.screen.get_size()
+    full_width, full_height = sys.screen.get_size()
     cell_width = int(full_width / nb_columns)
     cell_height = int(full_height / nb_rows)
 
@@ -17,6 +17,6 @@ def grid(
             x = int(col * cell_width)
             y = int(row * cell_height)
             position = (x, y)
-            with system.sub_screen(position, cell_width, cell_height):
+            with sys.sub_screen(position, cell_width, cell_height):
                 cell(row, col)
