+++ vsl/typing.py
@@ -1,9 +1,9 @@
-from vsl import system
+from vsl import sys
 
 
 def color(_color):
     if isinstance(_color, float) or isinstance(_color, int):
-        _color = system.etc.color_picker(_color)
+        _color = sys.etc.color_picker(_color)
 
     _color = (
         int(_color[0]),
